pyproject.toml
[project]
name = "parcha-strategy"
version = "0.1.0"
description = ""
authors = [
    {name = "AJ Asver",email = "aj@parcha.ai"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
